Split the "HttpUtils" into "HttpConnUtils" & "ConnSupportUtils".
